// styles/styles.d.ts
declare const globalStyles: {
  container: object;
  header: object;
  title: object;
  info: object;
  infoAlignedRight: object;
  data_box: object;
  fullBox: object;
  label: object;
  row: object;
  small: object;
  unit: object;
  rowAlignedBottom?: object; // Add this if you created it
  notifis: object;
  rowLabel: object;
  time: object;
  home:object;
  map:object;
};

export default globalStyles;